Update Maven configuration and enhance announcement features
The pom.xml is updated by adding dynamic versioning and removing unnecessary profiles. A build-and-upload pipeline is also introduced for GitHub Actions. Several additions to improve message announcements are made, such as enabling a 'chat' mode in messages, addition of an 'Announce' command, and creating a prefix configuration in PropagandaConfig.