Add support for multiple announcement types in AnnounceCommand
Updated the system to incorporate different types of announcements in the 'AnnounceCommand'. Users can now specify the type of announcements they want to send, such as chat, title, and action bar announcements. Support for these changes involves the creation of 'AnnouncementType' enumeration and the renaming of 'MessageType.java' to 'AnnouncementType.java'. The 'CommandManager' has also been adjusted to handle the new announcement types.